+++ application/views/cadastros/formapagamento-lista.php
@@ -1,10 +1,12 @@
-
+<?php $perfil_id = $this->session->userdata('perfil_id'); ?>
 <div class="content"> <!-- Inicio da DIV content -->
-    <div class="bt_link_new">
-        <a href="<?php echo base_url() ?>cadastros/formapagamento/carregarformapagamento/0">
-            Novo Plano
-        </a>
-    </div>
+    <?php if($perfil_id != 10){?>
+        <div class="bt_link_new">
+            <a href="<?php echo base_url() ?>cadastros/formapagamento/carregarformapagamento/0">
+                Novo Plano
+            </a>
+        </div>
+    <?php }?>
     <div id="accordion">
         <h3 class="singular"><a href="#">Manter Planos</a></h3>
         <div>
@@ -42,23 +44,35 @@
                         <tr >
                                 <td class="<?php echo $estilo_linha; ?>"><?= $item->nome; ?></td>
 
-                                <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                <td class="<?php echo $estilo_linha; ?>" width="70px;">  
+                                            <?php if($perfil_id != 10){?>
                                     <a href="<?= base_url() ?>cadastros/formapagamento/carregarformapagamento/<?= $item->forma_pagamento_id ?>">Editar</a>
+                                            <?php }?>
                                 </td>
-                                <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                <td class="<?php echo $estilo_linha; ?>" width="70px;">  
+                                      <?php if($perfil_id != 10){?>
                                     <a href="<?= base_url() ?>cadastros/formapagamento/carregarformapagamentocarencia/<?= $item->forma_pagamento_id ?>">Carência</a>
+                                     <?php }?>
                                 </td>
-                                <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                <td class="<?php echo $estilo_linha; ?>" width="70px;">   
+                                      <?php if($perfil_id != 10){?>
                                     <a href="<?= base_url() ?>cadastros/formapagamento/carregarformapagamentocomissao/<?= $item->forma_pagamento_id ?>">Comissão</a>
+                                    <?php }?>
                                 </td>
-                                <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                <td class="<?php echo $estilo_linha; ?>" width="70px;"> 
+                                      <?php if($perfil_id != 10){?>
                                     <a href="<?= base_url() ?>cadastros/formapagamento/anexararquivos/<?= $item->forma_pagamento_id ?>">Arquivos</a>
+                                     <?php }?>
                                 </td>
-                                <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
+                                <td class="<?php echo $estilo_linha; ?>" width="70px;"> 
+                                      <?php if($perfil_id != 10){?>
                                     <a   href="<?= base_url() ?>cadastros/formapagamento/cadastrarprocedimentoplano/<?= $item->forma_pagamento_id ?>">Procedimentos </a>
+                                     <?php }?>
                                 </td> 
-                                <td class="<?php echo $estilo_linha; ?>" width="70px;">                                  
-                                    <a onclick="javascript: return confirm('Deseja realmente exlcuir esse Forma?');" href="<?= base_url() ?>cadastros/formapagamento/excluir/<?= $item->forma_pagamento_id ?>">Excluir</a>
+                                <td class="<?php echo $estilo_linha; ?>" width="70px;">  
+                                    <?php if($perfil_id != 10){?>
+                                      <a onclick="javascript: return confirm('Deseja realmente exlcuir esse Forma?');" href="<?= base_url() ?>cadastros/formapagamento/excluir/<?= $item->forma_pagamento_id ?>">Excluir</a>
+                                    <?php }?>
                                 </td>
